Prepare for Mediawiki 1.39 and up (cont)
- Convert removed hook ParserBeforeStrip to ParserBeforeInternalParse

Note that this is not ideal when we want to tweak HTML because this
hook only can access to text fragments.

But moving to a later Hook (able to access to near-final HTML) can
have some severe implications, losing caching and other stuff.

So, while imperfect (we are removing a warning message, but not the
box containing it) it should be enough as far as it only happens
when a real-username page is going to be created.

There are more explanations in the onParserBeforeInternalParse()
function of the extension.

Closes #2